Moved CollectionRecord into user.py
